+++ optimlol_api/persistence/dataProviders/statsDataProvider.js
@@ -8,32 +8,6 @@ module.exports = function() {
 	var _logger = null;
 	var _performanceCalculator = require('../../common/performanceCalculator');
 
-	var _prepareStats = function(stats) {
-		if (stats.data) {
-			var optimlolChampions = [];
-			stats.data.champions.forEach(function(champion) {
-				var optimlolChampion = {};
-				optimlolChampion.id = champion.id;
-				optimlolChampion.wins = champion.stats.totalSessionsWon;
-				optimlolChampion.losses = champion.stats.totalSessionsLost;
-				optimlolChampion.kills = champion.stats.totalChampionKills;
-				optimlolChampion.deaths = champion.stats.totalDeathsPerSession;
-				optimlolChampion.assists = champion.stats.totalAssists;
-				if (champion.stats.totalDeathsPerSession) {
-					optimlolChampion.kda = (champion.stats.totalChampionKills + champion.stats.totalAssists) / champion.stats.totalDeathsPerSession;
-				} else {
-					optimlolChampion.kda = (champion.stats.totalChampionKills + champion.stats.totalAssists);
-				}
-				optimlolChampion.gamesPlayed = champion.stats.totalSessionsPlayed;
-				optimlolChampion.performance = _performanceCalculator.calculate(champion.stats.totalSessionsWon, champion.stats.totalSessionsLost);
-				optimlolChampions.push(optimlolChampion);
-			});
-
-			stats.data.champions = optimlolChampions;
-		}
-
-		return stats;
-	};
 
 	var _getStatsApi = function(region, summonerId, deferred) {
 		var statsPath = region + "/" + _apiVersion + "/stats/by-summoner/" + summonerId + "/ranked";
@@ -41,11 +15,11 @@ module.exports = function() {
 			.then(function(statsResult) {
 				_mongoCache.set('stats', {region: region, summonerId: summonerId}, statsResult)
 					.then(function() {
-						deferred.resolve(_prepareStats(statsResult));
+						deferred.resolve(statsResult);
 					})
 					.fail(function(error) {
 						_logger.warn("Some failure when setting stats cache", error);
-						deferred.resolve(_prepareStats(statsResult));
+						deferred.resolve(statsResult);
 					})
 			})
 			.fail(function(error) {
@@ -59,7 +33,7 @@ module.exports = function() {
 			.then(function(cacheStatsResult) {
 				if (cacheStatsResult.isExpired === false) {
 					_logger.debug("Using cached stats.");
-					deferred.resolve(_prepareStats(cacheStatsResult));
+					deferred.resolve(cacheStatsResult);
 				} else {
 					_getStatsApi(region, summonerId, deferred);
 				}
